refactor: add support for gr4vy id, remove duplicate code and old comments
