+++ .env.example
@@ -1,1 +1,2 @@
 PRIVATE_KEY=
+GR4VY_ID=
