+++ sdk/transactions.test.ts
@@ -1,24 +1,10 @@
-import fs from 'fs'
-import path from 'path'
 import { TransactionCaptureRequest } from '../model/transactionCaptureRequest'
 import { TransactionPaymentMethodRequest } from '../model/transactionPaymentMethodRequest'
 import { TransactionRefundRequest } from '../model/transactionRefundRequest'
 import { TransactionRequest } from '../model/transactionRequest'
-import Client from './client'
+import { getTestClient } from './helpers'
 
-let key
-if (process.env.PRIVATE_KEY) {
-  key = process.env.PRIVATE_KEY
-} else {
-  const my_path = path.resolve(__dirname, './private_key.pem')
-  key = String(fs.readFileSync(my_path))
-}
-
-const client = new Client({
-  gr4vyId: 'spider',
-  environment: 'sandbox',
-  privateKey: key,
-})
+const client = getTestClient()
 
 const AMOUNT = 1299
 let transactionId
@@ -27,7 +13,7 @@ jest.setTimeout(30000)
 
 const transactionRequest = new TransactionRequest()
 transactionRequest.amount = AMOUNT
-transactionRequest.currency = 'GBP'
+transactionRequest.currency = 'USD'
 transactionRequest.metadata = { source: 'node_sdk' }
 transactionRequest.paymentMethod = new TransactionPaymentMethodRequest()
 transactionRequest.paymentMethod.method =
